pyproject.toml
[project]
name = "football-scout-ai"
version = "0.1.0"
description = "Football Scout Deep Research Agent"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.1.0",
    "langchain-community>=0.2.0",
    "langchain-core>=0.2.0",
    "anthropic>=0.28.0",
    "tavily-python>=0.3.0",
    "soccerdata>=1.0.0",
    "aiosqlite>=0.20.0",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "httpx>=0.27.0",
    "sse-starlette>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]